requirements.txt
Django==5.2.7
djangorestframework==3.15.2
django-cors-headers==4.3.1
whitenoise==6.6.0
gunicorn==21.2.0